requirements.txt
streamlit
python-dotenv
streamlit-cookies-manager